Extract getModuleParts
